pv: vhe: Emulate trap for EL12 register accesses in EL1
On VHE systems with E2H bit set, a hypervisor accesses real EL0/EL1
registers by using EL02/EL12 registers.  On v8.3, accesses to EL02/EL12
in EL1 trap to EL2.

Since we work on v8.0, to make this trap happen, we paravirtualize a
guest hypervisor's access to EL12 to trap to EL2.

For EL02 accessors, which are all timer register accesses, they should
trap as well. But, for now, let the hypervisor access them natively.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>