pv: vhe: Do not trap on el2 reg accesses via el1 accessor
The current kvm code accesses el1 sys regs instead of el2 sys regs if
VHE is available.

If we run KVM hypervisor in EL1, and VHE is available, then
the KVM hypervisor will use EL1 accessors and we don't need to trap on it.
(On non-VHE systems, KVM accesses EL2 regs, so we had to trap.)

Let's not trap on any register accesses using read_sysreg_el2().

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>