KVM: arm64: Revive live_lrs
This is a partial revert of 00dafa0fcfe9fb1d863f08dc45d6f05ac9505d46.

This commit fixes the problem that interrupts are not delivered to
nested VMs in a timely manner. (e.g. delivering an IPI takes up to 40
sec.) This is because used_lrs variable was not updated when creating
the shadow lr. For example, even though the in-memory shadow lr has pending
entries, if used_lrs is zero, then the host hypervisor does not write
anything to the real list registers.

To fix this problem, We interate the whole lr regisisters since we can't
assume any specific way the guest hypervisor writes to the list
registers.

TODO: This is unoptimizing the non-nested KVM. Find a better way.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>