KVM: arm/arm64: Introduce data structure for rmap
We've been clearing the whole shadow stage-2 page tables in mmu notifier
handlers. This is because we are only given hva, but not nested VM's
physical address (L2 IPA), with which we can clear the exact mappings in
the shadow stage-2 page table.

This patch introduces data structure to manage mappings from hva to L2
IPA, which we call the reverse mappings or simply rmap. We begin with
having a kvm_rmap_head structure for each guest page frame (i.e. gfn).
The structure has information to locate a pte (or pmd if it's a huge
page mapping) in a shadow stage-2 page table corresponding to the given
hva. When multiple L2 IPAs are mapped to the same hva, the structure
points to a rmap_list_desc structure which can hold multiple L2 IPAs.

Signed-off-by: Jintack Lim <jintack.lim@linaro.org>