KVM: arm/arm64: Create rmap entries
When we create a mapping in the shadow stage-2 page table, we also
create an rmap entry.

This patch becomes very invasive because we need three information: L1
IPA, whether this mmu is a nested mmu, a pointer to rmap cache to create
a new descriptor if necessary. This ended up adding parameters to
stage2_set_pmd_huge() and stage2_set_pte(), which makes them look too
complex.

Signed-off-by: Jintack Lim <jintack.lim@linaro.org>