pv: Sanitize exception level in spsr_el2
On VHE systems, when there's an exception while running in the kernel,
then the exception level is EL2, which is saved in SPSR_EL2.

When a VHE guest hypervisor (i.e. kernel) is running in EL1 and when it
gets an exception, then the exception level saved in SPSR_EL1 is EL1.
This should be (virtual) EL2.

Therefore, we sanitize the exception level from EL1 to EL2 if this is
the case.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>